WP-415: sampler side of the commit for enyo-ilib samples
Add samples of how to parse and format person names, and how to parse and format mailing addresses. This is for stories WP-415 and WP-416. This change was actually done by Pugalendhi Ganesan and Guruprasad KN in LGSI, but they do not have push permissions, so I am pushing it for them.

Conflicts:
	source/package.js